+++ deltasigma/_ds_optzeros.py
@@ -16,6 +16,8 @@
 """Module providing the ds_optzeros() function
 """
 
+from __future__ import division
+
 import numpy as np
 from numpy import sqrt
 
@@ -78,7 +80,7 @@ def ds_optzeros(n, opt=1):
     """
     opt = int(opt)    
     if opt == 0:
-        optZeros = np.zeros((np.ceil(n/2.), ))
+        optZeros = np.zeros((int(np.ceil(n/2.)), ))
     else:
         optZeros = _oznopt[n][opt]
